+++ src/app/routes/reservations.ts
@@ -194,7 +194,9 @@ reservationsRouter.post(
                         object: { id: (<cinerino.factory.order.IReservation>acceptedOffer.itemOffered).id },
                         instrument: { token },
                         location: { identifier: (typeof locationIdentifier === 'string') ? locationIdentifier : undefined }
-                    })({ action: new cinerino.repository.Action(mongoose.connection) });
+                    })(
+                        // { action: new cinerino.repository.Action(mongoose.connection) }
+                    );
 
                     // 指定があれば、アクションIDをレスポンスに含める
                     if (includesActionId && typeof chevreUseAction?.id === 'string') {
@@ -229,9 +231,11 @@ function useReservation(params: {
         identifier?: string;
     };
 }) {
-    return async (repos: {
-        action: cinerino.repository.Action;
-    }) => {
+    return async (
+        // repos: {
+        //     action: cinerino.repository.Action;
+        // }
+    ) => {
         // 予約検索
         const reservationService = new cinerino.chevre.service.Reservation({
             endpoint: cinerino.credentials.chevre.endpoint,
@@ -243,15 +247,15 @@ function useReservation(params: {
 
         // 入場
         // 予約使用アクションを追加
-        const actionAttributes: cinerino.factory.action.IAttributes<cinerino.factory.actionType.UseAction, any, any> = {
-            project: { typeOf: cinerino.factory.chevre.organizationType.Project, id: params.project.id },
-            typeOf: cinerino.factory.actionType.UseAction,
-            agent: params.agent,
-            instrument: params.instrument,
-            object: [reservation]
-            // purpose: params.purpose
-        };
-        const action = await repos.action.start(actionAttributes);
+        // const actionAttributes: cinerino.factory.action.IAttributes<cinerino.factory.actionType.UseAction, any, any> = {
+        //     project: { typeOf: cinerino.factory.chevre.organizationType.Project, id: params.project.id },
+        //     typeOf: cinerino.factory.actionType.UseAction,
+        //     agent: params.agent,
+        //     instrument: params.instrument,
+        //     object: [reservation]
+        //     // purpose: params.purpose
+        // };
+        // const action = await repos.action.start(actionAttributes);
         let chevreUseAction: { id: string } | undefined;
 
         try {
@@ -266,17 +270,17 @@ function useReservation(params: {
             }
         } catch (error) {
             // actionにエラー結果を追加
-            try {
-                const actionError = { ...error, message: error.message, name: error.name };
-                await repos.action.giveUp({ typeOf: actionAttributes.typeOf, id: action.id, error: actionError });
-            } catch (__) {
-                // 失敗したら仕方ない
-            }
+            // try {
+            //     const actionError = { ...error, message: error.message, name: error.name };
+            //     await repos.action.giveUp({ typeOf: actionAttributes.typeOf, id: action.id, error: actionError });
+            // } catch (__) {
+            //     // 失敗したら仕方ない
+            // }
 
             throw error;
         }
 
-        await repos.action.complete({ typeOf: action.typeOf, id: action.id, result: {} });
+        // await repos.action.complete({ typeOf: action.typeOf, id: action.id, result: {} });
 
         return chevreUseAction;
     };