+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 全プロジェクトに許可されたアプリケーションクライアント管理を追加
 - 注文作成後の通知に注文トークンを付加
 - Chevre決済中止処理のエラーハンドリングを調整
+- 予約使用レスポンスにアクションIDを含めることができるように調整
 
 ### Deprecated