+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 予約使用レスポンスにアクションIDを含めることができるように調整
 - 予約に対する使用アクションをChevreで実装
 - 予約使用時にagent.identifierを指定できるように調整
+- Chevre予約の追加特性からpaymentSeatIndex,csvCode,transactionを削除
 
 ### Removed